% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPMOD.R
\name{NPMOD}
\alias{NPMOD}
\title{Non-Parametric Module}
\usage{
NPMOD()
}
\value{
NPMOD is a graphic interface
}
\description{
Non-Parametric Module (NPMOD) is a graphical interface dedicated to the testing of non-parametric data for educational purposes.

You can learn more about this package at:
http://www.uv.mx/personal/nehuerta/npmod/
}
\details{
This graphic interface was developed to provide techniques of non-parametric statistics for educational
purposes of easy management. On the other hand, this module contains a list of 17 non-parametric techniques that
yields results and graphics obtained from data.

Note: NPMOD is free software and comes with ABSOLUTELY NO WARRANTY.
}
\examples{
\dontrun{
##Install package
library(NPMOD)
##call the package
NPMOD()
}

}
\references{
Alan T. Arnholt. BSDA: Basic Statistics and Data Analysis, 2012.
R package version 1.01.

M Castillo, Arturo Ojeda, Mario Miguel Arturo Castillo, and Mario
Miguel Ojeda. Principios de estadistica no parametrica. Number 519.
54 C3P7. 1994.

Yosef Cohen and Jeremiah Y Cohen. Statistics and Data with R: An
applied approach through examples. John Wiley & Sons, 2008.

Jean Dickinson Gibbons and Subhabrata Chakraborti. Nonparametric
statistical inference. Springer, 2011.

Castor Guisande Gonzalez. Tratamiento de datos. Ediciones Diaz de
Santos, 2006.

Juergen Gross and Uwe Ligges. nortest: Tests for Normality, 2015.
R package version 1.0-4.

Maxime Herve. RVAideMemoire: Diverse Basic Statistical and Graphical
Functions, 2017. R package version 0.9-63.

Hans-Michael Kaltenbach. A concise guide to statistics. Springer Science
& Business Media, 2011.

Paul H Kvam and Brani Vidakovic. Nonparametric statistics with applications
to science and engineering, volume 653. John Wiley & Sons, 2007.

J Lopez. Introduccion al analisis de datos con r y r commander en psicologia
y educacion, 2012.

John H McDonald. Handbook of biological statistics, volume 2. Sparky House
Publishing Baltimore, MD, 2009.

John Verzani. Based on the iwidgets code of Simon Urbanek, suggestions by Simon Urbanek,
Philippe Grosjean, and Michael Lawrence. gWidgets: gWidgets API for building toolkit-independent,
interactive GUIs, 2014. R package version 0.0-54.

C Perez. Tecnicas estadisticas con spss 12. Aplicaciones al analisis de datos.
madrid. editorial pearson prentice hall, 2005.

R Core Team. R: A Language and Environment for Statistical Computing. R Foundation
for Statistical Computing, Vienna, Austria, 2017.

Myra L Samuels, Jeffrey A Witmer, and Andrew Schaffner. Statistics for the life sciences.
Pearson education, 2012.

Petr Savicky. pspearman: Spearman’s rank correlation test, 2014. R package version 0.3-0.

Venkatraman E. Seshan. clinfun: Clinical Trial Design and Data Analysis Functions, 2016.
R package version 1.0.13.

John Verzani. gWidgetstcltk: Toolkit implementation of gWidgets for tcltk package, 2014.
R package version 0.0-55.

Dennis D Mendenhall Wackerly, William Scheaffer, Richard L Romo Munoz, et al.
Estadistica matematica con aplicaciones. Number 519.5 W3. 2010.

Ronald M Weiers. Introduccion a la estadıstica para negocios.
Cengage Learning Editores, 2006.

Hadley Wickham. readxl: Read Excel Files, 2016. R package version 0.1.1.

Dirk Eddelbuettel with contributions by Antoine Lucas, Jarek Tuszynski,
Henrik Bengtsson, Simon Urbanek, Mario Frasca, Bryan Lewis, Murray Stokely,
Hannes Muehleisen, Duncan Murdoch, Jim Hester, Wush Wu, Qiang Kou, Thierry
Onkelinx, Michel Lang, and Viliam Simko. digest: Create Compact Hash Digests
of R Objects, 2017. R package version 0.6.12.
}
\author{
Juan Pablo Paulino-Arce, Nery Sofia Huerta-Pacheco, Victor Manuel Aguirre-Torres
}
